requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
python-docx==1.1.0
google-generativeai==0.3.2
openai>=1.40.0
pydantic==2.5.0
slowapi==0.1.9
python-dotenv==1.0.0

